Initial commit
PfsFileSelector is a widget that displays a window for file selection.
It is meant to be used with portal implementations but can be used for
other purposes too.

It contains a bottom pane and a directory view (PfsDirView). The view is
filled with grid items (PfsGridItem).

The widget can be used from C as well, see `examples/example.c`

Signed-off-by: Guido Günther <agx@sigxcpu.org>